+++ ttp/__init__.py
@@ -3,4 +3,4 @@
 
 """TTP - Transparent Tor Proxy."""
 
-__version__ = "0.4.5"
+__version__ = "0.4.6"